RenderTextures can get additional attachments
Diffuse and Specular part of the deferred shading tech use a
rendertexture to be put in. the specular part is an additional
attachment. Both get applied to the final composition output of the
scenerendering.

Next: Apply all lighttypes to the new rendering tech.
More later: reconfigure shadowmapping to fit deferred shading tech